refactor(shared): reorganize shared folder into subfolders by type
