--- src/shared/base.repository.ts
@@ -1,7 +1,0 @@
-export abstract class BaseRepository<T> {
-  abstract findAll(): Promise<T[]>;
-  abstract findById(id: number): Promise<T | null>;
-  abstract create(entity: T): Promise<T>;
-  abstract update(id: number, entity: Partial<T>): Promise<T>;
-  abstract delete(id: number): Promise<void>;
-}
