+++ src/shared/models/http-exception-response.model.ts
